again

+++ public/views/manage/controllers/admin-edit-ctrl.js
@@ -1,36 +1,43 @@
 var app = angular.module('lds-report');
 
-app.controller('adminEditCtrl', function($scope,userService){
+app.controller('adminEditCtrl', function($scope, userService, nzSwal){
 
-// var getAllAdmin = function(){
-// 	userService.getAll()
-// 	.then(function(response){
-// 		$scope.administrators = response.data;
-// 	})
-// }	
-
-// getAllAdmin();
 
 $scope.deleteAdmin = function(){
-		console.log($scope.admin.id)
-	userService.delete($scope.admin.id)
-	.then(function(res, err){
-		if(res === ""){
-			
-		}
-	
+		nzSwal({  
+			title : "Are you sure?",   
+			text: "This action cannot be undone!",   
+			type: "warning",   
+			showCancelButton: true,   
+			confirmButtonColor: "#DD6B55",   
+			confirmButtonText: "Yes, delete it!",   
+			closeOnConfirm: false 
+		})
+		 .then(function(){   
+		 	nzSwal("Deleted!", "Admin has been deleted.", "success"); 
+		 	userService.delete($scope.admin.id)
+		})
+		 .catch(function(){
+		 	nzSwal('Cancelled');
+		})
 
-	});
+		
 }
 
+
+
+
 $scope.editAdmin = function(){
-	console.log($scope.admin);
 	userService.edit($scope.admin.id, $scope.admin)
 	.then(function(res){
-		
-		
-		//todo User feedback was the change saved successfully
-	});
+	if(res.data.message){
+			$scope.error = res.data.message;
+		} else {
+			swal("Success!", "Admin profile has been modified!", "success")
+
+		}
+
+	})
 }
